Made message handlers for the backend, so server.ts is a little cleaned up/refactored
